pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "printd"
version = "0.1.0"
description = "An HTTP API for ESC/POS thermal receipt printers — USB, network, or serial."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Hans Fraiponts" }]
requires-python = ">=3.11"
keywords = ["escpos", "thermal-printer", "receipt", "fastapi", "pos"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: FastAPI",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Printing",
  "Topic :: System :: Hardware",
]
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "python-escpos>=3.1",
  "Pillow>=10.4",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "pytest-cov>=5.0",
  "httpx>=0.27",
  "ruff>=0.6",
]

[project.scripts]
printd = "printd.__main__:main"
printd-cli = "printd.cli:main"

[project.urls]
Homepage = "https://github.com/HansF/printd"
Repository = "https://github.com/HansF/printd"
Issues = "https://github.com/HansF/printd/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/printd"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"src/printd/app.py" = ["B008"]  # FastAPI File(...) default

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
